+++ cwism/wism/Pathing/IPathingStrategy.cs
@@ -6,6 +6,6 @@ namespace BranallyGames.Wism.Pathing
 {
     public interface IPathingStrategy
     {
-        void FindShortestRoute(Tile[,] map, Army source, Tile target, out IList<Tile> fastestRoute, out int distance);
+        void FindShortestRoute(Tile[,] map, Army source, Tile target, out IList<Tile> fastestRoute, out float distance);
     }
 }
